[REM3-413] Properly treat the scenario where a connection is closed pre auth closed.
See RemoteConnection.handlePreAuthCloseRequest to fully understand the code path. The remote connection write listener needs to have shutdown writes invoked to be aware that the channel is closed, or else there is no connection between the connection and the XNIO channels layer when the channel is closed at ServerConnectionOpenListener.Initial listener. To complete the fix, the very same remote connection write listener needs to skip the write attempt if the channel is closed, which will prevent an uneeded ClosedChannelException.

Signed-off-by: Flavia Rainone <frainone@redhat.com>